+++ CHANGELOG.md
@@ -32,6 +32,8 @@ This change log follows the conventions of
   groups of lights.
 - A new `confetti` effect which assigns random colors (and optionally
   aim points) to groups of lights at intervals.
+- A new `pinstripes` effect which can alternate stripes of color
+  across fixtures.
 - Incoming MIDI System Exclusive messages can now be received and
   delivered to handlers.
 - The identity of grid controllers is verified before binding to them,
@@ -83,6 +85,8 @@ This change log follows the conventions of
 - The effect overflow indicators in the Push text area were not
   properly disappearing when enough effects ended to render them no
   longer needed until the user pressed Shift to try to scroll.
+- Under some circumstances the Push mapping could crash when there was
+  no cue associated with an effect.
 - All MIDI event handler functions are now called in a context which
   properly recovers from exceptions at the level of that individual
   handler, so other handlers will not be affected.